test(sdk): Move integration tests
