+++ crates/matrix-sdk/src/client/builder.rs
@@ -343,12 +343,12 @@ impl ClientBuilder {
 }
 
 fn homeserver_from_name(server_name: &ServerName) -> String {
-    #[cfg(not(test))]
+    #[cfg(not(any(test, feature = "__test")))]
     return format!("https://{}", server_name);
 
     // Mockito only knows how to test http endpoints:
     // https://github.com/lipanski/mockito/issues/127
-    #[cfg(test)]
+    #[cfg(any(test, feature = "__test"))]
     return format!("http://{}", server_name);
 }
