+++ crates/matrix-sdk/tests/integration/room/mod.rs
@@ -1,0 +1,3 @@
+mod common;
+mod joined;
+mod left;
